+++ src/components/heatmap/HeatmapSubscriber.js
@@ -9,7 +9,6 @@ import {
   useCellsData,
   useCellSetsData,
   useExpressionMatrixData,
-  useGeneSelection,
 } from '../data-hooks';
 import { getCellColors } from '../interpolate-colors';
 import {
@@ -21,6 +20,7 @@ import {
 } from '../../app/state/coordination';
 import Heatmap from './Heatmap';
 import HeatmapTooltipSubscriber from './HeatmapTooltipSubscriber';
+import HeatmapOptions from './HeatmapOptions';
 
 const HEATMAP_DATA_TYPES = ['cells', 'cell-sets', 'expression-matrix'];
 
@@ -73,9 +73,9 @@ export default function HeatmapSubscriber(props) {
     geneHighlight,
     cellSetSelection,
     cellSetColor,
-    cellColorEncoding,
     additionalCellSets,
-    geneExpressionColormapRange: heatmapControls,
+    geneExpressionColormap,
+    geneExpressionColormapRange,
   }, {
     setHeatmapZoomX: setZoomX,
     setHeatmapZoomY: setZoomY,
@@ -85,7 +85,8 @@ export default function HeatmapSubscriber(props) {
     setGeneHighlight,
     setCellSetSelection,
     setCellSetColor,
-    setGeneExpressionColormapRange: setHeatmapControls,
+    setGeneExpressionColormapRange,
+    setGeneExpressionColormap,
   }] = useCoordination(COMPONENT_COORDINATION_TYPES.heatmap, coordinationScopes);
 
   const observationsTitle = capitalize(observationsPluralLabel);
@@ -115,9 +116,6 @@ export default function HeatmapSubscriber(props) {
   const [expressionMatrix] = useExpressionMatrixData(
     loaders, dataset, setItemIsReady, addUrl, true,
   );
-  const [expressionData] = useGeneSelection(
-    loaders, dataset, setItemIsReady, false, geneSelection, setItemIsNotReady,
-  );
   const [cellSets] = useCellSetsData(
     loaders, dataset, setItemIsReady, addUrl, false,
     { setCellSetSelection, setCellSetColor },
@@ -129,15 +127,16 @@ export default function HeatmapSubscriber(props) {
   ), [cellSets, additionalCellSets]);
 
   const cellColors = useMemo(() => getCellColors({
-    cellColorEncoding,
-    expressionData: expressionData && expressionData[0],
+    // Only show cell set selection on heatmap labels.
+    cellColorEncoding: 'cellSetSelection',
     geneSelection,
     cellSets: mergedCellSets,
     cellSetSelection,
     cellSetColor,
     expressionDataAttrs: expressionMatrix,
-  }), [cellColorEncoding, mergedCellSets, geneSelection,
-    cellSetColor, cellSetSelection, expressionData, expressionMatrix]);
+    theme,
+  }), [mergedCellSets, geneSelection, theme,
+    cellSetColor, cellSetSelection, expressionMatrix]);
 
   const getCellInfo = useCallback((cellId) => {
     if (cellId) {
@@ -171,6 +170,14 @@ export default function HeatmapSubscriber(props) {
       theme={theme}
       removeGridComponent={removeGridComponent}
       isReady={isReady && !isRendering}
+      options={(
+        <HeatmapOptions
+          geneExpressionColormap={geneExpressionColormap}
+          setGeneExpressionColormap={setGeneExpressionColormap}
+          geneExpressionColormapRange={geneExpressionColormapRange}
+          setGeneExpressionColormapRange={setGeneExpressionColormapRange}
+        />
+      )}
     >
       <Heatmap
         ref={deckRef}
@@ -182,14 +189,15 @@ export default function HeatmapSubscriber(props) {
           setTargetX(target[0]);
           setTargetY(target[1]);
         }}
-        heatmapControls={heatmapControls}
-        setHeatmapControls={setHeatmapControls}
+        colormapRange={geneExpressionColormapRange}
+        setColormapRange={setGeneExpressionColormapRange}
         height={height}
         width={width}
         theme={theme}
         uuid={uuid}
         expressionMatrix={expressionMatrix}
         cellColors={cellColors}
+        colormap={geneExpressionColormap}
         setIsRendering={setIsRendering}
         setCellHighlight={setCellHighlight}
         setGeneHighlight={setGeneHighlight}