+++ src/components/sets/CellSetSizesPlotSubscriber.js
@@ -75,9 +75,9 @@ export default function CellSetSizesPlotSubscriber(props) {
   // From the cell sets hierarchy and the list of selected cell sets,
   // generate the array of set sizes data points for the bar plot.
   const data = useMemo(() => (mergedCellSets && cellSetSelection && cellSetColor
-    ? treeToSetSizesBySetNames(mergedCellSets, cellSetSelection, cellSetColor)
+    ? treeToSetSizesBySetNames(mergedCellSets, cellSetSelection, cellSetColor, theme)
     : []
-  ), [mergedCellSets, cellSetSelection, cellSetColor]);
+  ), [mergedCellSets, cellSetSelection, cellSetColor, theme]);
 
   return (
     <TitleInfo