+++ src/components/sets/TreeNode.js
@@ -9,7 +9,7 @@ import PopoverMenu from './PopoverMenu';
 import HelpTooltip from './HelpTooltip';
 import { callbackOnKeyPress, colorArrayToString, getLevelTooltipText } from './utils';
 import { ReactComponent as MenuSVG } from '../../assets/menu.svg';
-import { DEFAULT_COLOR } from '../utils';
+import { getDefaultColor } from '../utils';
 
 
 /**
@@ -109,6 +109,7 @@ function NamedSetNodeStatic(props) {
     size,
     datatype,
     editable,
+    theme,
   } = props;
   const shouldCheckNextLevel = (level === 0 && !expanded);
   const nextLevelToCheck = (
@@ -151,7 +152,7 @@ function NamedSetNodeStatic(props) {
       {popoverMenuConfig.length > 0 ? (
         <PopoverMenu
           menuConfig={makeNodeViewMenuConfig(props)}
-          color={level > 0 && editable ? (color || DEFAULT_COLOR) : null}
+          color={level > 0 && editable ? (color || getDefaultColor(theme)) : null}
           setColor={c => onNodeSetColor(path, c)}
         >
           <MenuSVG className="node-menu-icon" />