+++ src/components/spatial/Spatial.js
@@ -1,19 +1,19 @@
 /* eslint-disable */
 import React, { forwardRef } from 'react';
 import isEqual from 'lodash/isEqual';
-import { ScatterplotLayer, PolygonLayer, COORDINATE_SYSTEM } from 'deck.gl';
-import { PointCloudLayer } from '@deck.gl/layers'; // eslint-disable-line import/no-extraneous-dependencies
-import { DataFilterExtension } from '@deck.gl/extensions'; // eslint-disable-line import/no-extraneous-dependencies
+import { COORDINATE_SYSTEM } from '@deck.gl/core'; // eslint-disable-line import/no-extraneous-dependencies
+import { PolygonLayer, PointCloudLayer } from '@deck.gl/layers'; // eslint-disable-line import/no-extraneous-dependencies
 import { Matrix4 } from 'math.gl';
 import { ScaleBarLayer, MultiscaleImageLayer } from '@hms-dbmi/viv';
-import { SelectablePolygonLayer, getSelectionLayers } from '../../layers';
-import { cellLayerDefaultProps, PALETTE, DEFAULT_COLOR } from '../utils';
+import { getSelectionLayers } from '../../layers';
+import { cellLayerDefaultProps, getDefaultColor } from '../utils';
 import { getSourceFromLoader } from '../../utils';
 import { square, getLayerLoaderTuple, renderSubBitmaskLayers } from './utils';
 import AbstractSpatialOrScatterplot from '../shared-spatial-scatterplot/AbstractSpatialOrScatterplot';
 import {
   createCellsQuadTree,
 } from '../shared-spatial-scatterplot/quadtree';
+import { ScaledExpressionExtension } from '../../layer-extensions';
 
 const CELLS_LAYER_ID = 'cells-layer';
 const MOLECULES_LAYER_ID = 'molecules-layer';
@@ -25,14 +25,21 @@ const makeDefaultGetCellPolygon = radius => (cellEntry) => {
   const cell = cellEntry[1];
   return cell.poly?.length ? cell.poly : square(cell.xy[0], cell.xy[1], radius);
 };
-const makeDefaultGetCellColors = cellColors => cellEntry => (
-  cellColors && cellColors.get(cellEntry[0])
-) || DEFAULT_COLOR;
-const makeDefaultGetCellIsSelected = cellSelection => cellEntry => (
-  cellSelection
-    ? cellSelection.includes(cellEntry[0])
-    : true // If nothing is selected, everything is selected.
-);
+const makeDefaultGetCellColors = (cellColors, theme) => (cellEntry) => {
+  const [r, g, b, a] = (cellColors && cellColors.get(cellEntry[0])) || getDefaultColor(theme);
+  return [r, g, b, 255 * (a || 1)];
+};
+const makeDefaultGetCellIsSelected = (cellSelection) => {
+  if (cellSelection) {
+    // For performance, convert the Array to a Set instance.
+    // Set.has() is faster than Array.includes().
+    const cellSelectionSet = new Set(cellSelection);
+    return cellEntry => (
+      cellSelectionSet.has(cellEntry[0]) ? 1.0 : 0.0
+    );
+  }
+  return () => 0.0;
+};
 
 /**
  * React component which expresses the spatial relationships between cells and molecules.
@@ -65,6 +72,7 @@ const makeDefaultGetCellIsSelected = cellSelection => cellEntry => (
  * @param {function} props.updateViewInfo Handler for DeckGL viewport updates,
  * used when rendering tooltips and crosshairs.
  * @param {function} props.onCellClick Getter function for cell layer onClick.
+ * @param {string} props.theme "light" or "dark" for the vitessce theme
  */
 class Spatial extends AbstractSpatialOrScatterplot {
   constructor(props) {
@@ -93,6 +101,12 @@ class Spatial extends AbstractSpatialOrScatterplot {
       width: 2048,
     };
     this.color = { ...this.randomColorData };
+    this.expression = {
+      data: new Uint8Array(2048 * 2048),
+      // This buffer should be able to hold colors for 2048 x 2048 ~ 4 million cells.
+      height: 2048,
+      width: 2048,
+    };
 
     // Initialize data and layers.
     this.onUpdateCellsData();
@@ -110,6 +124,7 @@ class Spatial extends AbstractSpatialOrScatterplot {
     } = layerDef;
     const { cellsEntries } = this;
     const {
+      theme,
       cellFilter,
       cellSelection,
       setCellHighlight,
@@ -118,11 +133,15 @@ class Spatial extends AbstractSpatialOrScatterplot {
         cellsEntries.length === cellSelection.length ? null : cellSelection,
       ),
       cellColors,
-      getCellColor = makeDefaultGetCellColors(cellColors),
+      getCellColor = makeDefaultGetCellColors(cellColors, theme),
       getCellPolygon = makeDefaultGetCellPolygon(radius),
       onCellClick,
       lineWidthScale = 10,
       lineWidthMaxPixels = 2,
+      geneExpressionColormapRange,
+      cellColorEncoding,
+      getExpressionValue,
+      geneExpressionColormap,
     } = this.props;
     const filteredCellsEntries = (cellFilter
       ? cellsEntries.filter(cellEntry => cellFilter.includes(cellEntry[0]))
@@ -132,14 +151,17 @@ class Spatial extends AbstractSpatialOrScatterplot {
     // so we need to flip it for rendering tooltips.
     const flipYTooltip = true;
 
-    return new SelectablePolygonLayer({
+    return new PolygonLayer({
       id: CELLS_LAYER_ID,
       backgroundColor: [0, 0, 0],
       isSelected: getCellIsSelected,
       getPolygon: getCellPolygon,
       updateTriggers: {
-        getFillColor: [opacity],
         getLineWidth: [stroked],
+        isSelected: cellSelection,
+        getExpressionValue,
+        getFillColor: [opacity, cellColorEncoding, cellSelection, cellColors],
+        getLineColor: [cellColorEncoding, cellSelection, cellColors],
       },
       getFillColor: (cellEntry) => {
         const color = getCellColor(cellEntry);
@@ -160,9 +182,18 @@ class Spatial extends AbstractSpatialOrScatterplot {
       getLineWidth: stroked ? 1 : 0,
       lineWidthScale,
       lineWidthMaxPixels,
+      getExpressionValue,
+      extensions: [new ScaledExpressionExtension({ instanced: false })],
+      colorScaleLo: geneExpressionColormapRange[0],
+      colorScaleHi: geneExpressionColormapRange[1],
+      isExpressionMode: cellColorEncoding === 'geneSelection',
+      colormap: geneExpressionColormap,
       ...cellLayerDefaultProps(
-        filteredCellsEntries, undefined, setCellHighlight,
-        setComponentHover, flipYTooltip,
+        filteredCellsEntries,
+        undefined,
+        setCellHighlight,
+        setComponentHover,
+        flipYTooltip,
       ),
     });
   }
@@ -328,6 +359,11 @@ class Spatial extends AbstractSpatialOrScatterplot {
       modelMatrix = new Matrix4(layerDef.modelMatrix);
     }
     if (rawLayerDef.type === 'bitmask') {
+      const {
+        geneExpressionColormap,
+        geneExpressionColormapRange = [0.0, 1.0],
+        cellColorEncoding,
+      } = this.props;
       return new MultiscaleImageLayer({
         // `bitmask` is used by the AbstractSpatialOrScatterplot
         // https://github.com/vitessce/vitessce/pull/927/files#diff-9cab35a2ca0c5b6d9754b177810d25079a30ca91efa062d5795181360bc3ff2cR111
@@ -344,10 +380,15 @@ class Spatial extends AbstractSpatialOrScatterplot {
         // has to do with the fact that we don't have it in the `defaultProps`,
         // could be wrong though.
         cellColorData: this.color.data,
-        cellColorHeight: this.color.height,
-        cellColorWidth: this.color.width,
+        cellTexHeight: this.color.height,
+        cellTexWidth: this.color.width,
         excludeBackground: true,
         onViewportLoad: layerProps.callback,
+        colorScaleLo: geneExpressionColormapRange[0],
+        colorScaleHi: geneExpressionColormapRange[1],
+        isExpressionMode: cellColorEncoding === 'geneSelection',
+        colormap: geneExpressionColormap,
+        expressionData: this.expression.data,
       });
     }
     const [Layer, layerLoader] = getLayerLoaderTuple(
@@ -433,7 +474,8 @@ class Spatial extends AbstractSpatialOrScatterplot {
     const { size } = this.props.cellColors;
     if (typeof size === 'number') {
       const cellIds = this.props.cellColors.keys();
-      color.data = new Uint8Array(color.height * color.width * 3).fill(128);
+      color.data = new Uint8Array(color.height * color.width * 3)
+        .fill(getDefaultColor(this.props.theme)[0]);
       // 0th cell id is the empty space of the image i.e black color.
       color.data[0] = 0;
       color.data[1] = 0;
@@ -449,6 +491,14 @@ class Spatial extends AbstractSpatialOrScatterplot {
       }
     }
     this.color = color;
+  }
+
+  onUpdateExpressionData() {
+    const { expressionData } = this.props;
+    if (expressionData[0]?.length) {
+      this.expression.data = new Uint8Array(this.expression.height * this.expression.width);
+      this.expression.data.set(expressionData[0]);
+    }
   }
 
   onUpdateMoleculesData() {
@@ -514,7 +564,14 @@ class Spatial extends AbstractSpatialOrScatterplot {
     }
 
     if ([
-      'layers', 'cells', 'cellFilter', 'cellSelection', 'cellColors',
+      'layers',
+      'cells',
+      'cellFilter',
+      'cellSelection',
+      'cellColors',
+      'geneExpressionColormapRange',
+      'cellColorEncoding',
+      'geneExpressionColormap',
     ].some(shallowDiff)) {
       // Cells layer props changed.
       this.onUpdateCellsLayer();
@@ -524,6 +581,12 @@ class Spatial extends AbstractSpatialOrScatterplot {
     if (['cellColors'].some(shallowDiff)) {
       // Cells Color layer props changed.
       this.onUpdateCellColors();
+      this.forceUpdate();
+    }
+
+    if (['expressionData'].some(shallowDiff)) {
+      // Expression data prop changed.
+      this.onUpdateExpressionData();
       this.forceUpdate();
     }
 
@@ -551,7 +614,18 @@ class Spatial extends AbstractSpatialOrScatterplot {
       this.forceUpdate();
     }
 
-    if (['layers', 'imageLayerLoaders', 'cellColors', 'cellHighlight', 'rasterLayersCallbacks'].some(shallowDiff)) {
+    if (
+      [
+        'layers',
+        'imageLayerLoaders',
+        'cellColors',
+        'cellHighlight',
+        'geneExpressionColormapRange',
+        'expressionData',
+        'rasterLayersCallbacks',
+        'geneExpressionColormap',
+      ].some(shallowDiff)
+    ) {
       // Image layers changed.
       this.onUpdateImages();
       this.forceUpdate();