+++ src/layers/bitmask-layer-shaders.js
@@ -1,4 +1,6 @@
-export const vs = `
+import glsl from 'glslify';
+
+export const vs = glsl`
 #define SHADER_NAME bitmask-layer-vertex-shader
 
 attribute vec2 texCoords;
@@ -20,9 +22,25 @@ void main(void) {
 }
 `;
 
-export const fs = `
+export const fs = glsl`
 #define SHADER_NAME bitmask-layer-fragment-shader
 precision highp float;
+
+#pragma glslify: plasma = require("glsl-colormap/plasma")
+#pragma glslify: viridis = require("glsl-colormap/viridis")
+#pragma glslify: greys = require("glsl-colormap/greys")
+#pragma glslify: magma = require("glsl-colormap/magma")
+#pragma glslify: jet = require("glsl-colormap/jet")
+#pragma glslify: bone = require("glsl-colormap/bone")
+#pragma glslify: copper = require("glsl-colormap/copper")
+#pragma glslify: density = require("glsl-colormap/density")
+#pragma glslify: inferno = require("glsl-colormap/inferno")
+#pragma glslify: cool = require("glsl-colormap/cool")
+#pragma glslify: hot = require("glsl-colormap/hot")
+#pragma glslify: spring = require("glsl-colormap/spring")
+#pragma glslify: summer = require("glsl-colormap/summer")
+#pragma glslify: autumn = require("glsl-colormap/autumn")
+#pragma glslify: winter = require("glsl-colormap/winter")
 
 // Data (mask) texture
 uniform sampler2D channel0;
@@ -40,6 +58,11 @@ uniform float hovered;
 // range
 uniform bool channelIsOn[6];
 
+// Expression mapping
+uniform vec2 uColorScaleRange;
+uniform bool uIsExpressionMode;
+uniform sampler2D expressionTex;
+
 // opacity
 uniform float opacity;
 
@@ -51,7 +74,9 @@ vec4 sampleAndGetColor(sampler2D dataTex, vec2 coord, bool isOn){
   // Colors are laid out corresponding to ids in row-major order in the texture.  So if width of the texture is 10, and you want ID 25,
   // you need coordinate (1, 4) (i.e 2 rows down, and 5 columns over indexed from 0 for a total of 25 units covered in row major order).
   vec2 colorTexCoord = vec2(mod(sampledData, colorTexWidth) / colorTexWidth, floor(sampledData / colorTexWidth) / (colorTexHeight - 1.));
-  vec4 sampledColor = vec4(texture(colorTex, colorTexCoord).rgb, 1.);
+  float expressionValue = texture(expressionTex, colorTexCoord).r / 255.;
+  float scaledExpressionValue = (expressionValue - uColorScaleRange[0]) / max(0.005, (uColorScaleRange[1] - uColorScaleRange[0]));
+  vec4 sampledColor = float(uIsExpressionMode) * COLORMAP_FUNC(clamp(scaledExpressionValue, 0.0, 1.0)) +  (1. - float(uIsExpressionMode)) * vec4(texture(colorTex, colorTexCoord).rgb, 1.);
   // Only return a color if the data is non-zero.
   return max(0., min(sampledData, 1.)) * float(isOn) * (sampledColor + hoveredColor);
 }